pricing/monitor/requirements.txt
azure-identity
azure-monitor-query